Remove the newline for the call function
If we add the newline in call arguments. Then the newline and leading
whitespaces will be treated as part of the arguments. The `words` and
`firstword` functions in GNU make 3.82 will treat newline as word separator
while GNU make 4.3 won't. This results the function "nwords` will report error
like: "first argument to `word' function must be greater than 0.  Stop."